+++ ziplime/finance/slippage/volatility_volume_share.py
@@ -1,5 +1,4 @@
 import math
-from decimal import Decimal
 
 from ziplime.finance.constants import ROOT_SYMBOL_TO_ETA, DEFAULT_ETA
 
@@ -43,8 +42,8 @@ class VolatilityVolumeShare(MarketImpactBase):
         # dictionary that always returns the same value.
         # NOTE: This dictionary does not handle unknown root symbols, so it may
         # be worth revisiting this behavior.
-        if isinstance(eta, (int, Decimal)):
-            self._eta = DummyMapping(Decimal(eta))
+        if isinstance(eta, (int, float)):
+            self._eta = DummyMapping(float(eta))
         else:
             # Eta is a dictionary. If the user's dictionary does not provide a
             # value for a certain contract, fall back on the pre-defined eta
