+++ ziplime/trading/entities/orders/stop_limit_order.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from ziplime.trading.entities.orders.order_request import OrderRequest
 from ziplime.trading.enums.order_side import OrderSide
 from ziplime.trading.enums.order_type import OrderType
@@ -7,8 +5,8 @@ from ziplime.trading.enums.order_type import OrderType
 
 class StopLimitOrder(OrderRequest):
 
-    def __init__(self, order_id: str, order_side: OrderSide, quantity: Decimal, limit_price: Decimal,
-               stop_price: Decimal):
+    def __init__(self, order_id: str, order_side: OrderSide, quantity: float, limit_price: float,
+               stop_price: float):
         super().__init__(order_id=order_id, order_type=OrderType.LIMIT)
         self.order_side = order_side
         self.quantity = quantity
